Move utility methods into Utils
