+++ src/uk/gov/defra/ffc/Utils.groovy
@@ -18,4 +18,22 @@ class Utils implements Serializable {
     def prUser = "${dbName}_$prSchema"
     return [prSchema, prUser]
   }
+
+  static def getRepoName(ctx) {
+    return ctx.scm.getUserRemoteConfigs()[0].getUrl().tokenize('/').last().split("\\.git")[0]
+  }
+
+  static def getMergedPrNo(ctx) {
+    def mergedPrNo = ctx.sh(returnStdout: true, script: "git log --pretty=oneline --abbrev-commit -1 | sed -n 's/.*(#\\([0-9]\\+\\)).*/\\1/p'").trim()
+    return mergedPrNo ? "pr$mergedPrNo" : ''
+  }
+
+  static def getRepoUrl(ctx) {
+    return ctx.sh(returnStdout: true, script: "git config --get remote.origin.url").trim()
+  }
+
+  static def getFolder(repoName) {
+    def folderArray = repoName.split('-')
+    return "${folderArray[0]}-${folderArray[1]}"
+  }
 }
