+++ vars/deploy.groovy
@@ -1,15 +1,12 @@
+import uk.gov.defra.ffc.Utils
+
 // public
 def trigger(jenkinsUrl, repoName, token, params) {
-  def folder = getFolder(repoName)
+  def folder = Utils.getFolder(repoName)
   def url = "$jenkinsUrl/job/$folder/job/$repoName-deploy/buildWithParameters?token=$token"
   params.each { param ->
     url = url + "\\&$param.key=$param.value"
   }
   echo "Triggering deployment for $url"
   sh(script: "curl -fk $url")
-}
-
-def getFolder(repoName) {
-  def folderArray = repoName.split('-')
-  return "${folderArray[0]}-${folderArray[1]}"
 }
